Extract builder types and add tests for update-flow-version (#542)
